Documentation generation for iOS SDKs
Summary:
Using sourcekitten output directly as a workaround to limitations with Jazzy in order to combine documentation for the ObjC and Swift interfaces of the iOS SDK.

Background:
Jazzy does not support mixed ObjC and Swift targets. Jazzy uses sourcekitten under the hood to generate the AST that it uses for generating docs. Sourcekitten can only generate docs for a Swift target or an ObjC target at one time. Hence Jazzy can only generate docs for either a Swift or an ObjC target in a single command.

Jazzy has a feature to generate documentation from a sourcekitten generated AST directly.

This means we can run sourcekitten twice, combine the output, and run Jazzy on that to generate mixed Swift and ObjC docs.

Reviewed By: dreamolight

Differential Revision: D17447656

fbshipit-source-id: de610128d5749cbad1022e493cd62780feb2d33e