Fixing dynamic schemes for Swift (#1061)
Summary:
Thanks for proposing a pull request!

To help us review the request, please complete the following:

- [x] sign [contributor license agreement](https://developers.facebook.com/opensource/cla)
- [x] I've ensured that all existing tests pass and added tests (when/where necessary)
- [x] I've updated the documentation (when/where necessary) and [Changelog](CHANGELOG.md) (when/where necessary)
- [x] I've added the proper label to this pull request (e.g. `bug` for bug fixes)

## Pull Request Details

Adds metadata indexer to dynamic schemes. Need to sort out how this got through CI earlier. Probably a gap that might cause this to happen in the future.
Pull Request resolved: https://github.com/facebook/facebook-objc-sdk/pull/1061

Test Plan: Should be able to build the dyanamic swift schemes in xcode (will need to add them as they are not currently part of the shared schemes)

Reviewed By: dreamolight

Differential Revision: D17666658

Pulled By: joesus

fbshipit-source-id: b44dcee0802fe7f164fa162dd6ee4bf2f6e30dbc