Adding new travis stage for combining artifacts of prior stages (#1063)
Summary:
Thanks for proposing a pull request!

To help us review the request, please complete the following:

- [x] sign [contributor license agreement](https://developers.facebook.com/opensource/cla)
- [x] I've ensured that all existing tests pass and added tests (when/where necessary)
- [x] I've updated the documentation (when/where necessary) and [Changelog](CHANGELOG.md) (when/where necessary)
- [x] I've added the proper label to this pull request (e.g. `bug` for bug fixes)

## Pull Request Details

Travis jobs run in parallel for each build stage. This ensures that the artifacts of prior stages will exist before they are combined and uploaded to the release.
Pull Request resolved: https://github.com/facebook/facebook-objc-sdk/pull/1063

Test Plan: Will have to watch the output of the next release but the changes are pretty straight-forward.

Reviewed By: dreamolight

Differential Revision: D17673442

Pulled By: joesus

fbshipit-source-id: 91c627d9ac57e36cb67435e3440e1931604b7b1b