Modularity Breaking BUCK
Summary:
Not positive why the `modular = True` flag was causing the error `redefinition of FBSDKCoreKit`. It seemed to be creating a module.modulemap file for each library associated with FBSDKCoreKit and declaring FBSDKCoreKit as the module name in each... ex: target FBSDKCoreKit_Basics was creating a module with the name FBSDKCoreKit.

Fixing by removing the `modular = True` flag which breaks Swift in unit tests. Removing those tests for now, adding task to reenable T55925870.

Differential Revision: D18216906

fbshipit-source-id: 2259f21f459221f62b3687f1a2bd01af2895b4f3