Making it easier to run unit tests (#1095)
Summary:
Thanks for proposing a pull request!

To help us review the request, please complete the following:

- [x] sign [contributor license agreement](https://developers.facebook.com/opensource/cla)
- [x] I've ensured that all existing tests pass and added tests (when/where necessary)
- [x] I've updated the documentation (when/where necessary) and [Changelog](CHANGELOG.md) (when/where necessary)
- [x] I've added the proper label to this pull request (e.g. `bug` for bug fixes)

## Pull Request Details

Before the unit tests were building the Swift Targets. This meant that trying to run tests on the shared dynamic schemes did not work.

It makes more sense to add unit tests later that are specifically targeting the added Swift functionality and that are associated with the Swift targets.

Once we drop support for Xcode 10.2 this distinction can be erased but for now it's a better experience to be able to easily run the tests for the target being worked on.
Pull Request resolved: https://github.com/facebook/facebook-ios-sdk/pull/1095

Test Plan:
Open workspace in Xcode. Run the unit tests for the schemes:
FBSDKCoreKit-Dynamic
FBSDKLoginKit-Dynamic
FBSDKShareKit-Dynamic

Reviewed By: jingping2015

Differential Revision: D18307413

Pulled By: joesus

fbshipit-source-id: c55829eae42ec350aee2fd54267aa1c20cdc6937