Updating FBSDKAutoAppLinkPresentationStyle's Swift name (#1099)
Summary:
Thanks for proposing a pull request!

To help us review the request, please complete the following:

- [x] sign [contributor license agreement](https://developers.facebook.com/opensource/cla)
- [x] I've ensured that all existing tests pass and added tests (when/where necessary)
- [x] I've updated the documentation (when/where necessary) and [Changelog](CHANGELOG.md) (when/where necessary)
- [x] I've added the proper label to this pull request (e.g. `bug` for bug fixes)

## Pull Request Details

It is not possible to Swift-namespace an ObjC enum into an ObjC protocol. This renames so that the symbol can be made available to Swift.
Pull Request resolved: https://github.com/facebook/facebook-ios-sdk/pull/1099

Test Plan:
Add a Swift package to a sample project and point to this branch `spm-patch-1`
In code, reference `AutoAppLinkPresentationStyle.auto`.
Should not throw an error.

Reviewed By: KylinChang

Differential Revision: D18383863

Pulled By: joesus

fbshipit-source-id: ac23fc14b6b8ddaf0bda3dc9bf40799b64201792