Symlink verification for public headers (#1115)
Summary:
Pull Request resolved: https://github.com/facebook/facebook-ios-sdk/pull/1115

SPM requires that C headers exist in an 'include' directory in order to be visible to consuming libraries.

This is similar to how CocoaPods' podspecs use the "public_header_files" argument.

In CocoaPods we can use a list of directories that have public headers. SPM is much more prescriptive in that it wants all of the public headers in one directory. We are using directory of symlinks to meet this requirement.

This Travis job should prevent us from mistakenly shipping new features that have not been exposed through this symlinking mechanism.

Reviewed By: jingping2015

Differential Revision: D18513231

fbshipit-source-id: 7cb8f8571500018812cd287e999b770e5e5a7caf