add back #import "FBSDKCoreKit+Internal.h" to solve swift package build error
Summary: Even though FBSDKGraphRequestPiggybackManager.h has contains `#import "FBSDKCoreKit+Internal.h"` in the file of `FBSDKGraphRequestConnection.m` and the project can build for coffeeShop and FBSDKCoreKitSwift. But if opening project with package.swift, we get some wired errors. So adding back the import and will investigate later why we get this issue

Differential Revision: D19276391

fbshipit-source-id: 31e07fdf1af711fb352f4226da997c0140e5a843