Adding ShareDialog stub for tvOS (#1175)
Summary:
Thanks for proposing a pull request!

To help us review the request, please complete the following:

- [x] sign [contributor license agreement](https://developers.facebook.com/opensource/cla)
- [x] I've ensured that all existing tests pass and added tests (when/where necessary)
- [x] I've updated the documentation (when/where necessary) and [Changelog](CHANGELOG.md) (when/where necessary)
- [x] I've added the proper label to this pull request (e.g. `bug` for bug fixes)

## Pull Request Details

This is sort of unfortunate. We are essentially stubbing classes so that we can compile Swift extensions that will never be run.

To support tvOS with SPM we need to use runtime availability checks in the Swift files. This means that even though the Swift extension of ShareDialog will never be run for tvOS targets, it still needs to be able to compile so we stub it out in the header. Eventually we will get away from this pattern entirely.
Pull Request resolved: https://github.com/facebook/facebook-ios-sdk/pull/1175

Test Plan: The SPM integration test caught this issue so a passing build should be enough to verify a fix.

Reviewed By: jingping2015

Differential Revision: D19330608

Pulled By: joesus

fbshipit-source-id: 8f0d9659e480c46afdb32ad2ce809d1bc813caf6