Removing Scrumptious. Updating README (#1190)
Summary:
Thanks for proposing a pull request!

To help us review the request, please complete the following:

- [x] sign [contributor license agreement](https://developers.facebook.com/opensource/cla)
- [x] I've ensured that all existing tests pass and added tests (when/where necessary)
- [x] I've updated the documentation (when/where necessary) and [Changelog](CHANGELOG.md) (when/where necessary)
- [x] I've added the proper label to this pull request (e.g. `bug` for bug fixes)

## Pull Request Details

Deletes sample project Scrumptious. Justification:
Does not build out of the box. Unclear how to install dependencies. Does not include a README. Has dead links to hosted content. Is unclear what it is trying to demonstrate. Promotes open graph sharing which is not encouraged. It was last updated last month to add some dependencies but before that it wasn’t worked on for three years.
Pull Request resolved: https://github.com/facebook/facebook-ios-sdk/pull/1190

Test Plan: N/A

Reviewed By: jingping2015

Differential Revision: D19504843

Pulled By: joesus

fbshipit-source-id: 5d9d28a04367e362d2f05c160ceda9f460fc4106