convert dictionary to NSData before write to file
Summary: Previously we call [NSDictionary writeToFile:atomically:] to save crash and method mapping as files on user device. Seems we were not able to control the content of dictionary and it caused crash `attempt to insert nil object from objects` https://fburl.com/tasks/j2wkcdzm. NSData is compatible with the invalid values, so this diff adds NSData/NSDictionary conversion.

Reviewed By: KylinChang

Differential Revision: D19544126

fbshipit-source-id: 4af57403ddcc67af871f5a21a37f49ea2c9d135c