Stabilize flaky static podspec linting (#1199)
Summary:
Thanks for proposing a pull request!

To help us review the request, please complete the following:

- [x] sign [contributor license agreement](https://developers.facebook.com/opensource/cla)
- [x] I've ensured that all existing tests pass and added tests (when/where necessary)
- [x] I've updated the documentation (when/where necessary) and [Changelog](CHANGELOG.md) (when/where necessary)
- [x] I've added the proper label to this pull request (e.g. `bug` for bug fixes)

## Pull Request Details

Pretty sure there is no reason to ever lint FBSDKCoreKit by itself since it is linted as a dependent podspec by all of the other podspecs.

That said, for the extra safety it seems we can still lint it if it does not include itself as a dependent spec.
Pull Request resolved: https://github.com/facebook/facebook-ios-sdk/pull/1199

Test Plan: Travis

Reviewed By: ZebingZong

Differential Revision: D19524637

Pulled By: joesus

fbshipit-source-id: 3bc6d68ed0c6a1f49f881fbd00869a5a8bb9300a