Adding graph_domain support to FBSDKAccessToken. (#1197)
Summary:
Pull Request resolved: https://github.com/facebook/facebook-ios-sdk/pull/1197

Access tokens can now be granted for specific domains.

This updates the SDK to automatically start using 'fb.gg' when the graph_domain field returns 'gaming'.

This graph_domain is stored with the FBSDKAccessToken and is reset during normal token clearing (Logout, Expiration etc).

A convenience initializer was added so that this is not a breaking change.

Reviewed By: joesus

Differential Revision: D19449489

fbshipit-source-id: e0c96ad299d0e6051c980a74ac002511fc631c09