Only reporting errors in Travis release job (#1209)
Summary:
Thanks for proposing a pull request!

To help us review the request, please complete the following:

- [x] sign [contributor license agreement](https://developers.facebook.com/opensource/cla)
- [x] I've ensured that all existing tests pass and added tests (when/where necessary)
- [x] I've updated the documentation (when/where necessary) and [Changelog](CHANGELOG.md) (when/where necessary)
- [x] I've added the proper label to this pull request (e.g. `bug` for bug fixes)

## Pull Request Details

Previous release's xcodebuild command drowned Travis' logs and caused the build to fail. This should prevent this on the next release.

Note: There is a possibility that Travis will time out due to a LACK of output. If this happens we may need to add back xcpretty or find some other way to truncate output. Tried adding the '-quiet' flag to xcodebuild but this did not quiet output enough.
Pull Request resolved: https://github.com/facebook/facebook-ios-sdk/pull/1209

Test Plan: run `sh scripts/run.sh release github` this will not  actually push anything to github. You should see that there is minimal output. Also you can change something to fail and make sure that there is output as expected.

Reviewed By: Nathaaaalie

Differential Revision: D19628541

Pulled By: joesus

fbshipit-source-id: c08d4b921fa5a1a9d42ce2e859f9c5837608cf8a