Updating Swift runtime check to work with SceneDelegate (#1215)
Summary:
Thanks for proposing a pull request!

To help us review the request, please complete the following:

- [x] sign [contributor license agreement](https://developers.facebook.com/opensource/cla)
- [x] I've ensured that all existing tests pass and added tests (when/where necessary)
- [x] I've updated the documentation (when/where necessary) and [Changelog](CHANGELOG.md) (when/where necessary)
- [x] I've added the proper label to this pull request (e.g. `bug` for bug fixes)

## Pull Request Details

For newer projects you may have seen: `FBSDKLog: Unable to find a valid UIWindow`. This is because there is code that checks for the existence of Swift. It does this check by seeing if the top view controller is namespaced (the module name-spacing implies Swift usage. This is not a perfect check but it is intended to give us a rough idea of how how many apps use Swift). Anyway, if a newer app is using SceneDelegate then the window is not available at the time we check this (AppDelegate's didFinishLaunching method). This delays the check by a second so that the window will be available.

**Why are we checking for the existence of Swift in apps that use the SDK?**
We would like to check for the existence of the Swift runtime to help make decisions around the direction of the SDK. For example, should we continue to support building with Xcode 10.2? What will the impact be of rewriting some classes / apis in Swift?
Pull Request resolved: https://github.com/facebook/facebook-ios-sdk/pull/1215

Test Plan: Create a new app in objective c and add a root view controller that uses Swift. Make sure the code is invoked to send an app event with the name: "fb_sdk_swift_runtime_check" and that the "is_using_swift" value is true.

Reviewed By: ZebingZong

Differential Revision: D19664079

Pulled By: joesus

fbshipit-source-id: ce2687ce31704c745e8d3d287bae1fe4594adbb7