Deleting Catalog sample app (#1233)
Summary:
Thanks for proposing a pull request!

To help us review the request, please complete the following:

- [x] sign [contributor license agreement](https://developers.facebook.com/opensource/cla)
- [x] I've ensured that all existing tests pass and added tests (when/where necessary)
- [x] I've updated the documentation (when/where necessary) and [Changelog](CHANGELOG.md) (when/where necessary)
- [x] I've added the proper label to this pull request (e.g. `bug` for bug fixes)

## Pull Request Details

There's nothing completely wrong with this sample but it should be split into smaller pieces that represent individual kits so that the samples are easier to consume. Additionally the dependencies should be dealt with in a way that requires less intervention from the developer. Currently it's unclear how to install dependencies and the project does not build out of the box, it has a readme but it is minimal.
Pull Request resolved: https://github.com/facebook/facebook-ios-sdk/pull/1233

Test Plan: N/A

Reviewed By: robtimp

Differential Revision: D19799495

Pulled By: joesus

fbshipit-source-id: fe8840005fbdd9b34d1cb018447be2732146220b