update FBSDKModelRuntime to support MTML with task parameter
Summary: If I understand correctly, compared to MTML and existing model, everything is the same except the name of last layer. So with existing model, if we pass the task as the name of "fc3" (model side it probably be "dense3"), everything should work as before. But in this diff, we are exposing a new API with task parameter so that we can support MTML with new weights file.

Differential Revision: D19874844

fbshipit-source-id: 3838b5c6d2670dc735e4af2fe2303a1322e701b9