update FBSDKModelRuntime to support MTML with task parameter (#1243)
Summary:
Pull Request resolved: https://github.com/facebook/facebook-ios-sdk/pull/1243

If I understand correctly, compared to MTML and existing model, everything is the same except the name of last layer. So with existing model, if we pass the task as the name of "fc3" (model side it probably be "dense3"), everything should work as before. But in this diff, we are exposing a new API with task parameter so that we can support MTML with new weights file.

Reviewed By: KylinChang

Differential Revision: D19874844

fbshipit-source-id: 46ede4284d8afea566258c6293a477c49f237821