move dispatch_once outside of ModelManager and add more tests
Summary: Follow up on D19945774, according to Apple's doc (https://developer.apple.com/documentation/dispatch/1447169-dispatch_once), This function is useful for initialization of global data (singletons) in an application. Always call this function before using or testing any variables that are initialized by the block. So move the dispatch_once to where the enable is called and add tests for MTML disabled case.

Differential Revision: D19987719

fbshipit-source-id: a4ca0eb601aeb6af2a6779412ee93eea6f6747cb