Deleting ShareIt Sample App (#1255)
Summary:
Thanks for proposing a pull request!

To help us review the request, please complete the following:

- [x] sign [contributor license agreement](https://developers.facebook.com/opensource/cla)
- [x] I've ensured that all existing tests pass and added tests (when/where necessary)
- [x] I've updated the documentation (when/where necessary) and [Changelog](CHANGELOG.md) (when/where necessary)
- [x] I've added the proper label to this pull request (e.g. `bug` for bug fixes)

## Pull Request Details

Deletes ShareIt sample app.

If you try and build this app you’ll see that it includes an AppID for an app in development mode. You’ll also notice that you cannot share because the sample content has been flagged as abusive. In short, this app provides zero value. It was last updated 8 months ago but before that it wasn’t worked on for two years. Additionally it has been replaced by the FacebookSharingSample app.
Pull Request resolved: https://github.com/facebook/facebook-ios-sdk/pull/1255

Test Plan: N/A

Reviewed By: Mxiim

Differential Revision: D20070798

Pulled By: joesus

fbshipit-source-id: 3181db8a3f47364ac24c072374ac0b824b83ca93