Sharing sample (#1236)
Summary:
Thanks for proposing a pull request!

To help us review the request, please complete the following:

- [x] sign [contributor license agreement](https://developers.facebook.com/opensource/cla)
- [x] I've ensured that all existing tests pass and added tests (when/where necessary)
- [x] I've updated the documentation (when/where necessary) and [Changelog](CHANGELOG.md) (when/where necessary)
- [x] I've added the proper label to this pull request (e.g. `bug` for bug fixes)

## Pull Request Details

Sample project to show how ShareKit works on a very basic level.
Pull Request resolved: https://github.com/facebook/facebook-ios-sdk/pull/1236

Test Plan:
Spent 30 minutes getting the current version to compile. Needed to run carthage build, update the paths to the build, needed to change several compiler flags, then when it was finally building, it instacrashed on an error about a missing image file. I got it to  build a few months ago and found that it includes an AppID for an app in development mode. You’ll also notice that you cannot share because the sample content has been flagged as abusive.

The new one, you open and it pulls down the dependency. Hit run. It  guides you through entering an app id, then you can share.

{F228404254}

{F228404251}

{F228404248}

{F228404250}

{F228404249}

{F228404255}

Reviewed By: jingping2015

Differential Revision: D19817133

Pulled By: joesus

fbshipit-source-id: 9dc332d1934e5a88f6fd1099f8bdda236fd7206a