Fix spm integration (#1272)
Summary:
Thanks for proposing a pull request!

To help us review the request, please complete the following:

- [x] sign [contributor license agreement](https://developers.facebook.com/opensource/cla)
- [x] I've ensured that all existing tests pass and added tests (when/where necessary)
- [x] I've updated the documentation (when/where necessary) and [Changelog](CHANGELOG.md) (when/where necessary)
- [x] I've added the proper label to this pull request (e.g. `bug` for bug fixes)

## Pull Request Details

I found there are two possible ways of setting up an integration test for SPM

First way would be to point to this repo. From there we can test the branch we're trying to merge in, or we can test master. We cannot point to the merge commit because Travis CI doesn’t actually push the merge commit to any branch (which is good normally but in this case we want it to). I opted to solve this by force-pushing the merge commit to a dummy branch and then testing against that.

The alternate way to set this up would be to point to the package with a file URL. This works locally but on both Travis CI and Github Workflows it results in an error that looks like:

```
xcodebuild: error: Could not resolve package dependencies:
An unknown error occurred. '/Users/runner/Library/Developer/Xcode/DerivedData/SmoketestSPM-bkrljbedpiwridakhqlllisyfeas/SourcePackages/checkouts/facebook-ios-sdk'
 exists and is not an empty directory (-4)
```
Pull Request resolved: https://github.com/facebook/facebook-ios-sdk/pull/1272

Test Plan: CI

Reviewed By: jingping2015

Differential Revision: D20283034

Pulled By: joesus

fbshipit-source-id: 246aadd6ee0ed456c13240aa37129374cd456e96