Adding MethodUsageMonitor (#1287)
Summary:
Thanks for proposing a pull request!

To help us review the request, please complete the following:

- [x] sign [contributor license agreement](https://developers.facebook.com/opensource/cla)
- [x] I've ensured that all existing tests pass and added tests (when/where necessary)
- [x] I've updated the documentation (when/where necessary) and [Changelog](CHANGELOG.md) (when/where necessary)
- [x] I've added the proper label to this pull request (e.g. `bug` for bug fixes)

## Pull Request Details

Adds FBSDKMethodUsageMonitor to be able to capture method usage in the SDK.
Ultimate goal is to be able to tell which methods are useful and which methods should maybe be deprecated.
Pull Request resolved: https://github.com/facebook/facebook-ios-sdk/pull/1287

Test Plan: Unit Tests

Reviewed By: dreamolight

Differential Revision: D20512740

Pulled By: joesus

fbshipit-source-id: 2fd000eaf3e3fe8495c10ea0f90a85cb88b50450