Point package integration to merge commit ref (#1327)
Summary:
Thanks for proposing a pull request!

To help us review the request, please complete the following:

- [x] sign [contributor license agreement](https://developers.facebook.com/opensource/cla)
- [x] I've ensured that all existing tests pass and added tests (when/where necessary)
- [x] I've updated the documentation (when/where necessary) and [Changelog](CHANGELOG.md) (when/where necessary)
- [x] I've added the proper label to this pull request (e.g. `bug` for bug fixes)

## Pull Request Details

The previous implementation would not work for forked repos because of how secrets are handled in Travis CI. This implementation should work for both and has the benefit of being simpler all around.
Pull Request resolved: https://github.com/facebook/facebook-ios-sdk/pull/1327

Test Plan: CI

Reviewed By: jingping2015

Differential Revision: D20928340

Pulled By: joesus

fbshipit-source-id: 3f87a6759428b46130f23385b8275e840423478e