Allow for building spm integration on either pr merge ref or master (#1336)
Summary:
Thanks for proposing a pull request!

To help us review the request, please complete the following:

- [x] sign [contributor license agreement](https://developers.facebook.com/opensource/cla)
- [x] I've ensured that all existing tests pass and added tests (when/where necessary)
- [x] I've updated the documentation (when/where necessary) and [Changelog](CHANGELOG.md) (when/where necessary)
- [x] I've added the proper label to this pull request (e.g. `bug` for bug fixes)

## Pull Request Details

We need to provide a valid branch to build the SPM  Integration project. In the case of PRs we want to use the merge ref provided by github. In the case that we're building master (after a merge) then we want to use the master branch.
Pull Request resolved: https://github.com/facebook/facebook-ios-sdk/pull/1336

Test Plan: CI

Reviewed By: KylinChang

Differential Revision: D20997655

Pulled By: joesus

fbshipit-source-id: 43f63e2f4a56f7f6b9a73ffa17a4e32221e3e12a