Fixing debugger crash in mixed CocoaPod - SPM projects (#1362)
Summary:
Thanks for proposing a pull request!

To help us review the request, please complete the following:

- [x] sign [contributor license agreement](https://developers.facebook.com/opensource/cla)
- [x] I've ensured that all existing tests pass and added tests (when/where necessary)
- [x] I've updated the documentation (when/where necessary) and [Changelog](CHANGELOG.md) (when/where necessary)
- [x] I've added the proper label to this pull request (e.g. `bug` for bug fixes)

## Pull Request Details

Fixes https://github.com/facebook/facebook-ios-sdk/issues/1360

This is pretty wild. To reproduce the issue create a new sample project with the following pods:

```
  pod 'FBSDKCoreKit'
  pod 'FBSDKLoginKit'
  pod 'FBSDKShareKit'
```

Run the project with a breakpoint.

Result:
Debugger works as expected.

Then add an arbitrary Swift Package. I used Alamofire to test this.
`https://github.com/alamofire/alamofire`

Run the project with a breakpoint.

Result:
Debugger is broken.

**Why this is happening:**
It seems that during compile time, the macro SWIFT_PACKAGE is not available but it is available at runtime.

**Solution**
Use a more specific macro to identify when we are building the Swift Package for the SDK and not rely on the system provided macro.
Pull Request resolved: https://github.com/facebook/facebook-ios-sdk/pull/1362

Test Plan: Follow repro instructions with the new code. Should be able to access symbols in the debugger.

Reviewed By: dreamolight

Differential Revision: D21302965

Pulled By: joesus

fbshipit-source-id: 5a97808f65b42808bbe7265465ef2b5d78e0e4a5