Adding release config check to CI (#1372)
Summary:
Thanks for proposing a pull request!

To help us review the request, please complete the following:

- [x] sign [contributor license agreement](https://developers.facebook.com/opensource/cla)
- [x] I've ensured that all existing tests pass and added tests (when/where necessary)
- [x] I've updated the documentation (when/where necessary) and [Changelog](CHANGELOG.md) (when/where necessary)
- [x] I've added the proper label to this pull request (e.g. `bug` for bug fixes)

## Pull Request Details

Travis CI does not support Catalina at this point. We need to make sure that it does not try to build Catalyst or it will fail.
Pull Request resolved: https://github.com/facebook/facebook-ios-sdk/pull/1372

Test Plan: Added CI test in order to test the portion of the release that was failing.

Reviewed By: KylinChang

Differential Revision: D21434764

Pulled By: joesus

fbshipit-source-id: 233a57fbbc45245c49a4eedb49b3cf41e7586068