Added unit tests for FBSDKRestrictiveDataFilterManager
Summary: Used **XCTAssert** to verify that the returned value from *processParameters(parameters, eventName)* was not Nil (this would mean parameters was empty or Nil) AND not equal to original params value (this would mean that *isRestrictiveEventFilterEnabled* was not set to YES from the *enable* function). As params are matched in custom events with the event name, the event name is not modified. **For security**, the eventName was checked to not be Nil. To do this, research was done on the OCMock iOS test framework, XCTestCase, FBSDKCoreKitTests, and ObjectiveC handling.

Reviewed By: KylinChang

Differential Revision: D21784075

fbshipit-source-id: d23690f3a3e3ff409f235736a59c62e77a319d62