Fix the issue of cannot opting out AAM
Summary:
Previously, we could not remove AAM data if we opt out some rules.
Now, we align with Android AAM and add an internalUserData to save AAM data. And we only send back the data of enabled AAM rules.

And we don't need to expose functions
```
+ (void)setHashData:(nullable NSString *)hashData
            forType:(FBSDKAppEventUserDataType)type;
+ (nullable NSString *)getHashedDataForType:(FBSDKAppEventUserDataType)type;
```

These two functions are only called in FBSDKMetadataIndexer previously.

And I renamed
`+ (nullable NSString *)getHashedDataForType:(FBSDKAppEventUserDataType)type;`
with
`+ (nullable NSString *)getInternalHashedDataForType:(FBSDKAppEventUserDataType)type`
for it's only used in FBSDKMetadataIndexer.

Reviewed By: dreamolight

Differential Revision: D21830610

fbshipit-source-id: b6840ef1554b9e9e0413f41277d3c64e358bbb5a