Use type safety linter to modify FBSDKCoreKit
Summary:
Ran:
`arc cast //xplat/tools/cAST:fbtidy //fbobjc/ios-sdk:FBSDKCoreKit -checks="-*,facebook-ios-sdk-utility-methods-check"`

To break it down.
"arc cast //xplat/tools/cAST:fbtidy //fbobjc/ios-sdk:FBSDKCoreKit" - run the linter on the FBSDKCoreKit target

"-checks="-*,facebook-ios-sdk-utility-methods-check" - with no checks except for the facebook-ios-sdk-utility-methods-check.

Manually effort was needed to import some headers.
Also needed to deal with call sites where we referenced members on collection return types. ex:
```
[dictionaryOfBools objectForKey:key].intValue
```
was translated to:
```
[FBSDKTypeUtility dictionary:dictionaryOfBools objectForKey:key ofType:NSObject.class].intValue
```
I manually adjusted to:
```
[[FBSDKTypeUtility dictionary:dictionaryOfBools objectForKey:key ofType:NSObject.class] intValue]
```

In the future we can make this better but the amount of time it would take to solve this particular issue makes it not worth it since there were only two places to change this.

Also was not sure how to exclude the test directory so instead just reverted changes in the test directory.

For obvious reasons also reverted the modded changes to FBSDKTypeUtility

Reviewed By: dreamolight

Differential Revision: D21789806

fbshipit-source-id: 39b2ab10e877d95b95bc44389688dc140a7a691d