Make pods modular
Summary:
Cocoapods that combine Swift and ObjC need to enable modules to be statically loaded.

So in the Podfile if you need to include `use_modular_headers!` if you do NOT include `use_frameworks!`.

This is a pain for users who don't care about how we handle modular headers in our SDK so we should make it easy for them to be able to load our libraries statically and still import them as modules.

More info:
http://blog.cocoapods.org/CocoaPods-1.5.0/
https://stackoverflow.com/questions/52910973/how-to-remove-use-frameworks-and-keep-using-swift-pods-on-an-objective-c-projec

Also updates ruby syntax to be more readable.

Differential Revision: D22048806

fbshipit-source-id: d0bfba23f7266733c65dfb6b7841b2aa35dc5af5