Add more tests to FBSDKAppEventsUtilityTests
Summary:
Add two tests to testParamsDictionary.

Test 1: set user data using FBSDKAppEvents, check whether the user data is correctly stored in parameter dictionary.
Result: If we don't set user data, a key-value pair (@"ud"-@"{}") is stored in the parameter dictionary. If we set the user data, the user data is correctly stored in the parameter dictionary and doesn't affect other data.
Test 2: set user id using FBSDKAppEvent, check whether the user id is correctly stored in parameter dictionary.
Result: If we set the user id, we can get the correct user id using the key "app_user_id" in the parameter dictionary. If we don't set the user id, there is no key-value pair for the user id in the parameter dictionary.

Implementation details:
Considering we set the user id in the setUp function, I tested the case with user id set at the beginning. For the case without user id set, I tested it in the end by clearing user id. For the first test, I tested the case without user data set first. Then tested the case with user data set and cleared the user data in order to remove effects on following tests.

Reviewed By: KylinChang

Differential Revision: D22335350

fbshipit-source-id: 023721eacb69b761d0902f574b2cfa744df66e50