Unit tests on FBSDKRestrictiveData
Summary:
In this diff, I write 8 unit tests on FBSDKRestrictiveData.
Test 1: a general case. We can easily iterate the server response and create FBSDKRestrictiveData.
Test 2: the structure of some fields is not expected. FBSDKRestrictiveData will correctly init to nil.
Test 3: the server adds some extra fields to the response. FBSDKRestrictiveData will be created correctly. It just ignores unknown fields.
Test 4: the field of restrictive_param is missing in the response. FBSDKRestrictiveData will set corresponding property to nil.
Test 4: the field of restrictive_param is missing in the response. FBSDKRestrictiveData will set corresponding property to nil.
Test 5: the field of deprecated_param is missing in the response. FBSDKRestrictiveData will set corresponding property to nil.
Test 6: the field of is_deprecated_event is missing in the response. FBSDKRestrictiveData will set the property of deprecatedEvent to NO.
Test 7: the field of restrictive_param is empty in the response, the restrictiveParams property will be empty as well.
Test 8: the field of deprecated_param is empty in the response, the deprecatedParams property will be empty as well.

Reviewed By: joesus

Differential Revision: D22363462

fbshipit-source-id: cd1d11746a87f36b3c25075a1136e7378f401d24