add FBSDKRestrictiveData class
Summary:
Context:
Restrictive data filtering is reading the event name and parameters being detected as sensitive from server side. Currently, we don't have structure validation for the response, so server side change will easily cause iOS app crash.

In this series of diffs:
1. create a FBSDKRestrictiveData class to check and parse server response.
2. write unit tests to test the FBSDKRestrictiveData class.
3. replace relevant code with FBSDKRestrictiveData in FBSDKRestrictiveDataFilterManager or other files.

In this diff,
I create a FBSDKRestrictiveData class to check and parse server response. If the structure of the server response is not expected, this class will init to nil. If the structure is valid, it just parses the response and assign values to corresponding properties.
It should be mentioned that the property deprecatedEvent will be set as NO if the is_deprecated_event field is missing in the server response.

Reviewed By: tianqibt

Differential Revision: D22363480

fbshipit-source-id: cd5663cf5aff0511ac57ec5bc3be9df7fc67291c