Xcode 12 related build fixes
Summary:
Stops linking UIKit binary for static FBSDKGamingServicesKit. Not 100% sure why this is failing the xcode-integration ci job but this might help. This is more consistent with how the other kits are linking (or not linking) UIKit and local testing showed no issue with removing it.

Additionally the carthage job is failing for tvos. This is likely due to the xcode 12 rollout. Including the fixes in here so we can fix the build as one diff.

allow-large-files
Also upgrading carthage binary 0.35.0 -> 0.36.0 to  to see if that's the reason for the command failure.

Differential Revision: D24196254

fbshipit-source-id: 51cb301b379fffa746deaea982d643d3c701f116