Fix Rsync Step
Summary:
I made some faulty assumptions about why I was getting an rsync warning so I gated it behind a file check. This led to a false positive where it appeared that the files were being synced when they were not.

This updated code ignores the output of the step when it fails since it will fail for GamingServicesKit (because it contains no Swift files at the moment).

Reviewed By: tianqibt

Differential Revision: D24339344

fbshipit-source-id: 59f03a1552c78aa39f86144ba99b6e8438e51c3f