Make Basics an actual dependency 2/n (#1526)
Summary:
Pull Request resolved: https://github.com/facebook/facebook-ios-sdk/pull/1526

## Context:

In terms of BUCK, CocoaPods and Swift Package Manager, CoreKitBasics is its own module.

Even in terms of xcodeproj it's its own target. This formalizes that dependency relationship.

## Why do this?

This allows us to avoid a lot of the messiness where we have something like:

```
#if defined FBSDKCOCOAPODS
 #import <FBSDKCoreKit/FBSDKCoreKit_Basics.h>
#elif defined BUCK
 #import <FBSDKCoreKit_Basics/FBSDKCoreKit_Basics.h>
#else
 #import "FBSDKCoreKit_Basics.h"
#endif
```
Now we can just do:
```
#import <FBSDKCoreKit_Basics/FBSDKCoreKit_Basics.h>
```
and it will "just work"

## This diff

Adds FBSDKCoreKit basics as a linked dependency to the other kits. This relationship is not new but it was previously hidden by having Basics source files added to the other CoreKit targets. This also fixes the CoreKit+Internal dependency to depend on the library instead of just importing the header.

Differential Revision: D24179749

fbshipit-source-id: 39387124566adf3690563823b2cbb1a51f658c9f