Make Basics an actual dependency 3/n
Summary:
## Context:

In terms of BUCK, CocoaPods and Swift Package Manager, CoreKitBasics is its own module.

Even in terms of xcodeproj it's its own target. This formalizes that dependency relationship.

## Why do this?

This allows us to avoid a lot of the messiness where we have something like:

```
#if defined FBSDKCOCOAPODS
 #import <FBSDKCoreKit/FBSDKCoreKit_Basics.h>
#elif defined BUCK
 #import <FBSDKCoreKit_Basics/FBSDKCoreKit_Basics.h>
#else
 #import "FBSDKCoreKit_Basics.h"
#endif
```
Now we can just do:
```
#import <FBSDKCoreKit_Basics/FBSDKCoreKit_Basics.h>
```
and it will "just work"

## This diff

Updates imports to work with Swift Package Manager. Looks like at this point it's unnecessary to change the Swift name in `FBSDKAppEventUserDataType` (see earlier versions of the diff for context). This might become necessary later when the Package is updated to support Swift but that's an unrelated change.

Reviewed By: dreamolight

Differential Revision: D24179748

fbshipit-source-id: 0a84f17b06af84af3c90ec3280a2b3f752d77354